+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "vscode-microbit-micropython" extension will be docum
 
 ## [0.1.x]
 
+- Added command to show image of micro:bit pinout (0.1.6)
 - Few more interface bugs + video tutorial link (0.1.5)
 - Fixing python dependency (0.1.4)
 - Improved README and fixed critical dependency python code (0.1.3)